requirements.txt
# Основные зависимости
playwright==1.40.0
aiofiles==23.2.1

# Дополнительные утилиты
requests==2.31.0
lxml==4.9.3
beautifulsoup4==4.12.2

# Для планировщика задач
schedule==1.2.0

# Для работы с конфигурацией
python-dotenv==1.0.0
pyyaml==6.0.1